requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-cors-headers
python-dotenv
gunicorn
gunicorn
django-jazzmin
pytest
pytest-django
xhtml2pdf
